requirements.txt
Django==1.8
Markdown==2.6.2
MySQL-python==1.2.5
Pillow==3.1.1
django-bootstrap3==6.2.2
django-crispy-forms==1.4.0
django-embed-video==0.11
django-filter==0.9.2
django-phonenumber-field==0.7.2
factory-boy==2.5.2
fake-factory==0.5.1
phonenumbers==7.0.4
reportlab==3.1.44
requests==2.6.0
wsgiref==0.1.2
xlrd==0.9.3
xlutils==1.7.1
xlwt==1.0.0
